Support persistent connections (keep-alive) in aserve client
Calls to do-http-request can now specify :keep-alive t and
do-http-request will return the socket connected to the server if
the server agrees with the request. That socket connection can be
used in subsequent calls to do-http-request.

Are there user visible changes in this commit?  yes

Is bootstrapping needed?   no

Are tests included for new features?  yes (in t-aserve.cl)

Tests run:  aserve/test/t-aserve

<release-note>
rfe9811 - the :keep-alive arguement to do-http-request will now
cause do-http-request to return a fifth value: a socket connected
to the server.  That fifth value will be nil if the server refuses
to keep the connection with the client open.   Even if the server
agrees to keeps the connection alive it will close it after some
period of inactivity.

do-http-request takes a :connection argument which is how the
socket returned by do-http-request should be passed to the
next do-http-request call.  It is ok if the value of
the :connection argument is nil or closed socket.  do-http-request
will notice that the socket is unusable and will create a new one.

</release-note>

<documentation>
rfe9811 has the documention changes
</documentation>

<documentation>

Change-Id: Ife1f8a9b7f0a4ba3377750c08534251e7a62aad6