Add support for Expect: 100-continue requests
aserve version: 1.2.70
The HTTP spec specifies that a client who does not want to send its
body before the server has confirmed that it is prepared to read it
can send an Expect: 100-continue header. The server then sends a
response with a 100 status code, after which the client sends its
body, and things continue as normal.

We have two concrete uses for this:

 1) Curl sends such a header before sending any big body. It will then
    wait for a while before sending its body anyway. This means that,
    by default, curl works with our server, but is much slower than it
    has to be for big requests. We have a work-around of specifying -H
    "Expect:", but it would be better if things just worked.

 2) The proxy used between the AG front and backend can't reliably use
    keep-alive when forwarding PUT or POST requests without this
    feature. If it forwards a request body, and the socket it was
    using turned out to be stale, it no longer has the body, and the
    request fails. It can send Expect: 100-continue as a way to test
    whether the server is still listening on the socket it has, before
    actually forwarding the body.

Change-Id: Ibd6c5b3c54883aac0681a577a02cf3d9a750a621