Support :connect method requests in do-http-request
Properly handle :connect method requests in do-http-request.

CONNECT requests differ slightly from all other request types in that
they MUST be passed through a proxy, and the value passed on the
Request-Line of the message is an Authority (host:port) instead of a
valid URI.

changes to keep-alive and skip-body semantics in do-http-request.
Other changes in make-client-http-request.

Are there user visible changes in this commit?  no

Are tests included for new features?  no new features

Tests run:  none

Change-Id: I89f5157782a573b5e96e9281fd7b49fa24e47490