rfe8891: support chunking with ssl
http/1.1 clients and servers prefer to send and receive
data using chunking.  This change allows Aserve to support
chunking in the client and server if the connection is ssl

Are there user visible changes in this commit?  yes

Is bootstrapping needed?   no

Are tests included for new features?  no, existing tests test new features

Tests run:  t-aserve

<release-note>
rfe8891: support chunking with ssl

In the past AllegoServe disabled chunking for an ssl server
and told users to use the http/1.0 protocol for clients when
ssl was being used.  Now those restrictions are gone.
The default for the server will be to use chunking in all
cases and clients can specify http/1.1 for requests over ssl.

</release-note>

<documentation>
just a note to the documentation editor to look for references
to chunking and ssl in the manual and remove the text.
</documentation>

Change-Id: Id2c2b2e34c1e49c8bfc14f0eb31122d21afa41cc