rfe10416: add compression
Support server gzip like compressing body before sending to client.
Support client gzip like uncompressing body as sent from server.

Are there user visible changes in this commit?  yes

Is bootstrapping needed?   no

Are tests included for new features?  yes

Tests run:  t-aserve

<release-note>
rfe10416 - enable compression

aserve can now compress responses on the fly if request.
aserve's client code can request a compressed transfer and can
uncompress the result.
See documentation for API changes.
Subsequent releases will increase aserve's capability to cover
static file transfers.

</release-note>

<documentation>
rfe10416 - enable compression

(do-http-request &key compress)
 takes a :compress argument (default nil) which if true causes
 the request to the server to include an 'Accept-Encoding: gzip'
 header line.  The server then may choose to return the result
 compressed to the client which will then uncompress it before
 returning the body from do-http-request.

(start &key compress)

The net.aserve:start function takes a :compress argument.  The value
of this argument is is stored in the enable-compression slot of
the wserver object.  The default value of this argument is
true if the deflate module was successfully loaded.

The enable-compression slot of the wserver object is a server wide
control on whether the server will compress a result before
sending it back.  If this slot has the value nil then under no
cicumstance will the server compress a value before returning it.
If the value is true then the server may or may not compress a
result before sending it to the client.

(publish-xxx  &key compress )

Each entity has a compress slot as well, set by the
compress argument of the various publish functions.  The default for
the :compress argument for all publish functions is true.

Furthermore the compute-strategy generic function that determines
how a result will be returned is the function that decides if
compression will be done.

At the present time (with changes planned for the future) the only
entity that can be compressed is the computed entity. In order
for compression to occur a number of things must be true

1. the enable-compression slot of the wserver object must be true
2. the compress slot of the entity must be true
3. the entity must be a computed entity (for now)
4. the client must state that it can accept gzip content-encoding

compression will work with any combination of chunking and/or ssl.

</documentation>

Change-Id: I683ff58a12447b8614263cafce5990ee491ed387