rfe10416/v1.3.0: return pre-compressed files and ssl protocol specification
aserve version 1.3.0

rfe10416 - If a compressed version of a static file is available and
the client is willing to accept a compressed encoding of the body
then send the compressed version in place of the actual file.

ssl - allow the user to specify the ssl protocols to be used in
the client and server ssl streams

Are there user visible changes in this commit?  yes

Is bootstrapping needed?   no

Are tests included for new features?  yes, tests for compressed files

Tests run:  t-aserve

<release-note>
rfe10416 - compression support

AllegroServe will now return a compressed version of a file object
in preference to an uncompressed version. See the documentation for
what must be true for this to occur.
</release-note>

<release-note>
ssl protocol selection

One can now specify the ssl protocol used by the server and
client code in AllegroServe.  The argument is
:ssl-method to net.aserve:start and net.aserve.client:do-http-request
The default is still to use any protocol but one might specify
:tlsv1 to ensure that only the most modern and secure protocol is used.
</release-note>

<documentation>
rfe10416 - compression support  documentation additons:

If a request comes in for a file AllegroServe may instead return
a compressed version of that file.  All of the following must
be true for this to occur
1. when the file was published in aserve with publish-file, the
  :compress argument had a true value.
  Note that if a directory is published with publish-directory then
  when a file is discovered by AllegroServe in that directory
  it is published by AllegroServe using publish-file and in this case
  the value of the :compress argument to publish-file is the same
  value as the original :compress argument to publish-directory.
2. The http client must have specified that it was willing to
   accept a compressed body using the Accept-Encoding header with
   a value that includes "gzip".  Modern web browsers will
   usually pass this Accept-Encoding header.
3. The compressed version of the file must exist and its name must end
   with a period followed by file type given in the
   compression-file-types slot
   of the *wserver* object.  Currently the default file type is "gz"
   for a gzip'ed file.
   The uncompressed file need not exist [although if the uncompressed
   version does not exist then the compressed version will NOT be
   uncompressed to satisfy a request that can only be satisified
   by the uncompressed version]
4. If the compressed and uncompressed versions of the file exist
   then the compressed version must be as young or younger than the
   uncompressed file.

Even if all the above is true, if the request comes in with
an If-Modified-Since header then AllegroServe will return a 304 Not Modified
if it deems that this file hasn't changed since the If-Modified-Since time.

Note that if all the conditions above are true and a compressed version
of a file is transmitted to the client, the caller of the client will never see
the compressed file.  The caller might ask the client to retrieve
http://server/foo.txt and the client will return the contents of
foo.txt to the caller even though foo.txt.gz was in fact transmitted
to the client by AllegroServe.
</documentation>

<documentation>
documentation for ssl changes:

add section:

SSL Protocol
The SSL protocol used for secure communcation has gone through
a sequence of revisions. The public revisions are SSL v2.0, SSL v3.0
and SSL v3.1.  There was a renaming as well so SSL v3.1 is officially
known as TLS v1.0.

When an SSL client connects to an SSL server they will communicate
using the most recent protocol that both support.

By default AllegroServe's SSL server and client declare that they
are willing to communicate using SSL v2.0, v3.0 or v3.1 (TLS v1.0).
If you wish to restrict the server or client to a particular
protocol you can pass the :ssl-method argument to net.aserve:start
or net.aserve.client:do-http-request.   At present the only
meaningful value you would want to pass is :tlsv1 meaning that
you only want to communicate using TLS v1.0 (SSL v3.1), the most modern
and secure of the protocols.

In the documention for net.aserve:start and net.aserve.client:do-http-request
add the :ssl-method argument and reference the SSL Protocol section
in the document

</documentation>

Change-Id: Ie424378d78f5d273b3028e35d12d7ac52c1b4e17