v1.3.1 - publish-directory can have a list of destination directories
publish-directory can accept a single destination directory or a
list of destination directories to be searched when a request is given.

Are there user visible changes in this commit?  yes

Is bootstrapping needed?   no

Are tests included for new features?  yes

Tests run:  t-aserve

<release-note>
v1.3.1 - publish direction can accept a list of destinations

When a request comes in publish-directory will search one
or more directories for the requested file.

</release-note>

<documentation>
publish-directory's :destination argument can now be a single
directory or a list of directories to search.  The first file
found of the requested name will be returned.
There are cases where the compressed version of a file is
returned in place of the uncompressed file.  In this case
the compressed and uncompressed versions of the file should
be in the same destination directory.  If they are in
different directories the last modified time comparison will
not be made so the compressed version may be returned even if
it is out of date.

</documentation>

Change-Id: I2b1f0b324fa67c812764ac04f75e2ee7c12bbb6d