v1.3.2 - delay sending headers for computed entities
For a computed entity send the response code and headers as late as possible
thus allowing one to handle an error in the computed entity function
by sending a different response, this time describing the error.

Are there user visible changes in this commit?  yes

Is bootstrapping needed?   no

Are tests included for new features?  yes

Tests run:  t-aserve

<release-note>
delay sending headers for computed entities

Errors can occur when running the computed entity function.
When an error occurs one would like to send back a different
response such as '500 - Internal Server Error'.  In order to
make this possible the http response code and the headers that
follow that are not immediately sent back to the server
when  with-http-body starts.   Instead they are held in
request-reply-stream and are only sent when the computed entity
function sends data to the stream.  Thus if you do computation
before sending any data you can avoid having any response
sent to the client until the computation finishes.

</release-note>

Change-Id: Iba2e494221820e0bbb72326888c45b56f3cc48f7