v1.3.3 add option to do hidden redirect to an index file in a directory
publish-directory arranges for a index file in a directory to be
returned when the uri specfies the directory name.  Normally aserve
returns a redirect to the index file which is then visible to the client.
There is now an option to simply return the contents of the index
file rather than do the redirect.

Are there user visible changes in this commit?  yes

Is bootstrapping needed?   no

Are tests included for new features?  yes

Tests run:  t-aserve

<release-note>
Changes to how directory entities return index files

First any uri from the client that refers to a directory on the disk
via a directory-entity will be changed (if necessary) to
end with a slash (/) via redirection.
Then the new hidden-index-redirect argument to publish-directory
will determine if a redirect is done to the index file or
if the contents of the index file are returned immediately.

</release-note>

<documentation>
 this change has been reflected in aserve.html (included with this
 commit).
</documentation>

Change-Id: I29ea111372fe0611ccd30960051262e122c1dfa4