bug20446. Fix log reporting of content-length with keep-alive streams.
If multiple requests are sent through a socket when using
keep-alive, log entries for each response report the
total bytes written out on the stream rather than
the actual bytes sent for just the response indicated in
the log entry.

Are there user visible changes in this commit?  no

Tests run:  none

Change-Id: I068591552132854653cdae4604b8b2c1d876d978