request-query cache checks external-format
v1.3.13 - spr38589 - request query caches the parse of the supplied form
and now remembers the external format of the parse so that if the external
format argument to request-query changes the parse will be redone.

Are there user visible changes in this commit?  yes/no

Is bootstrapping needed?   no

Are tests included for new features?  no tests for new features

Tests run:  t-aserve

Change-Id: Ia99c421f1bcafb46828c32779a587fcd3ad0879d