v1.3.14 - client sends all cookies on one line
rfe6265 says that a client must send all cookies on one line.  We've been
sending them on multiple lines for a long time and servers haven't
complained but now and this is starting to be a problem, likely due
to rfc6265 stating the one line requirement.

Are there user visible changes in this commit?  yes/no

Is bootstrapping needed?   no

Are tests included for new features?  no new features

Tests run:  t-aserve

Change-Id: Ibf135b107bce5223ed4586fe2c34857c1b2cb99d