fix for older allegro cl versions
... or any version for that matter without the recent patch for
with-output-to-buffer.

For logging bodies (and also to nicely accumulate multiple debug
messages of the same kind into a single one, we need
with-output-to-buffer that supports growable arrays which was first
available as a patch to acl 8.2. If that feature is not supported in
the lisp, we print an explanatory message to *error-output* and go on
without being able to log dynamically computed bodies (i.e. those that
aren't published files). Also, since accumulation doesn't work, one
can have several entries for the same logical thing.

Note that due to the nature of macrology, the decision whether to use
the fully featured or the slightly restricted version is made at
compile time.

Change-Id: Idd0863381c5c9f6ec6d23a2b757a932c5848353f