fix declaration in base64-decode
An incorrect array declaration was given but it's been there for a
long time so clearly the compiler had not used that declaration
to generate bad code.  However as a result the compiler did not
open code the array references.  base64-decode is not a critical
function so no performance difference are anticipated.

Change-Id: Ibe947e4ea706dbbfd04782c89eee7a49503c6d65