v1.3.17 - introduce allegroserve-error condition object
In order to add program recovery from errors detected by AllegroServe
we've introduced an allegroserve-error object which we'll begin
using in place of a generic lisp error.  Initially we're using
it in the client code to signal a timed out connection.

A documentation update will be filed.

This is a response to spr39616

Change-Id: I780e86f4e0521dc5a86bb7c4621e9810155e0e6c