bug21142: handle connection reset and aborted errors in the client
READ-SOCKET-LINE assumes that (READ-CHAR SOCKET NIL NIL) gets an eof
and returns NIL if the connection is severed. Windows can surprise us
with WSAECONNABORTED, and linux with ECONNRESET (and it's actually
documented). Other platforms are expected to vary in when/if these can
be triggered when reading from a socket, but it's safe to treat these
errors as eof for our purposes.

Update defpatch form (acl82: v17, acl90: v4).

<release-note>
bug21142 - handle connection reset and aborted errors in the client

Previously, net.client:do-http-request sometimes signaled connection
reset and connection aborted errors when the :connection argument was
used and the connection was severed (for example, the connection timed
out on the server and was closed). With this fix, connection reset and
aborted errors are treated as an eof: we retry the request with a new
connection which should make keep-alive operation more reliable.
</release-note>

Change-Id: I1ec4d1e6c3b3ea54c9f22bacf63d6080277c31ca