General support for 100-continue expect header handling.
Move automatic handling of 100 Continue responses until after an
entity is found and authorizers have been run, in order to provide
limited checking (401 and 404) of incoming requests before instructing
the client to deliver the body.

Make auto-handling configurable at the entity level. Add a
:will-handle-expect-continue keyword argument to each of the publish-*
routines (default: nil). If T, the entity will be reponsible for
sending a 100 response. If a request has an 'Expect: 100-cotinue'
header, set a flag in the request, retrievable via
request-has-continue-expectation accessor. If the body is read before
handling occurs, then a 100 reponse will be sent automatically if
necessary.

When aserve is used as a proxy, set will-handle-expect-continue to t
so that the header is forwarded.

If a 100-continue expect header is present in a request, have the
aserve client delay sending the body of the request until a 100
Continue response is seen.

Tests run: aserve test suite.

Change-Id: Ief44263192e18e4ef29699e50b619a160e5e1dcc