rfe12965: Allow full ssl config in aserve client/server.
Aserve exports only a subset of the ssl related keyword arguments. As
a result, it is not possible to fully customize a secure client or
server.

Add :ssl-args keyword to net.aserve:start,
net.aserve.client:do-http-request, and
net.aserve.client:make-http-client-request. As argument, it accepts a
plist of keyword options and values accepted by make-ssl-client-stream
or make-ssl-server-stream as pertains to the function being invoked.

If :ssl-args is specified all other ssl related keyword arguments are
ignored. This makes it possible to fully configure the ssl layer of an
http server or client, such as specifying the list of allowed ciphers,
and eliminates the need to patch AllegroServe when the underlying ssl
implementation changes.

Existing ssl-related keywords are retained for backwards
compatibility.

Change-Id: Ida0cdf195b89c3e806966b76f32df64e59415bc4