bug22596: Fix bug in do-http-request retry-on-timeout
Fix a bug in the retry-on-timeout logic in do-http-request. When
do-http-request attempts to retry a request due to a 408 response, it
recursively calls do-http-request again. After the retry and the
initial call to do-http-request was re-entered, it was attempting to
further process the response of the original timeout
response. Instead, the results of that recursive call should be
returned from the initial invocation of do-http-request where the
timeout occurred.

Tests added to test/t-aserve distinct from the other timeout related
tests.

Change-Id: If9ad3ab77bebc4a455281fcd6a811db120e90cb8