rfe13402: Use defaults of the underlying ssl module.
Remove outdated ssl defaults in both the client and server.
Instead, rely on the defaults of the underlying ssl module,

Fix a bug where passing an :ssl-args argument, and not an :ssl arg,
when starting a server would not result in an https server being
started. A server can now be started via

  :ssl "server.pem"

or

  :ssl-args '(:certificate "server.pem")

Updates the t-aserve testsuite to use the preferred :ssl-args when
performing https testing.

<release-note>
rfe13402: Use defaults of the underlying ssl module.

Remove the outdated :sslv23 defaults for using ssl in both the
Allegroserve client and server. Instead, rely on the defaults supplied
by the underlying ssl module.

</release-note>

<documentation>
see rfe13398
</documentation>

Change-Id: I44cbd848f1e5f9996c77400c22056b9e7f0322bc