bug22805: Have Aserve use the userinfo in request uri for basic-auth
If the request uri passed to an Allegroserve client function contains
user authentication information in the userinfo slot, use it. The
:basic-authorization keyword parameter takes precedence over the uri
userinfo field if both are specified.

Tests added to test/t-aserve

<release-note>
bug22805: Have Aserve use the userinfo in request uri for basic-auth

It is possible, though not recommended, to include user login
information in the userinfo part of a uri. For example, the following
uri includes login for the user "test" with a password of "friend":

  http://test:friend@examples.org/a/door

Previously, Allegroserve would ignore any userinfo included in the
uri, and only issue an Authorization header if the
:basic-authorization keyword were specified in the client call. Now,
if no explicit basic-authorization is passed to the client function,
Allegroserve will use the userinfo field in the request uri, if
present, to generate an authorization header for the request.

</release-note>

<documentation>
tNo doc update needed.
</documentation>

Change-Id: Iacc2b5fc04c4ec156bfded13b3b5021f53b700db