rfe13851: add MAX-LISTENERS arg to START
The LISTENERS argument of START determines only the number of HTTP
worker threads the server is going to start with. Under load the
server may create more workers which may lead to resource management
problems down the road.

For this reason, a new argument, MAX-LISTENERS was added to START.
When it is specified, it must be an integer denoting the hard limit on
the number of workers. The previous behavior corresponds to
MAX-LISTENERS being NIL (the default).

Change-Id: I1c25b08e8f30521519435bf763698e7562bf5a95