Fix processing of chunked requests sent over SSL
When processing a chunked HTTP request, the server used to rely
on the socket library to perform the dechunking operation.
The problem is that this is not currently implemented for SSL
sockets (see rfe8891). To remedy that, the server now uses its
own implementation of a dechunking stream. This is the same method
that is used to decode chunked replies in the client code.
In addition, trailing headers received with a chunked request are
now correctly processed (these were previously ignored, even for
non-SSL requests).

See bug23851 for more details regarding problems with chunked HTTPS
requests.

Changes:
    * main.cl: Version 1 3 39
    (get-request-body-retrieve): use a dechunking stream instead of
    turning on dechunking at the socket level. Process trailing
    headers.

    * makefile: added (require :tester) in the test.mp target.

    * packages.cl: Defpatch 10.0 (4)

    * test/t-aserve.cl (test-chunked-request): a new test for
    chunked requests with trailing headers.

Change-Id: I7ceea4567767aaf6687390228d5e09d6d0f13745