v1.3.44 - add :test-ssl argument to start
If an ssl server is being started and :test-ssl is true then
check before start returns if the certificates are valid and if
not signal an error.

test added: yes
tests run: yes

<release-note>
Added a :test-ssl argument to net.aserve:start enabling
a user to test immediately if they've suppplied a valid certificate
for an incoming http SSL request.
</release-note>

Documention changes made to aserve.html.

Change-Id: I8f2b1eeae9a756d0ad0a5c2b77b10e1708f6525a
Reviewed-on: https://gerrit.franz.com:9080/6607
Reviewed-by: John Foderaro <jkf@franz.com>
Reviewed-by: Kevin Layer <layer@franz.com>
Tested-by: Kevin Layer <layer@franz.com>