rfe15225: Add :no-keep-alive strategy
%Testing: aserve/test/t-aserve.cl

AllegroServe was sending one of the headers "Connection: Keep-Alive"
or "Connection: Close" with every reply.  The Websocket protocol adds
the header "Connection: Upgrade" in the reply to a Websocket connection
request.  The presence of two "Connection" headers was causing some
Microsoft browsers to fail Websocket connections to the ACL server.
By using this new strategy, the ACL Websocket server can suppress the
second (legal, but unnecessary) header.

Change-Id: Icdf6fadc589336298e2a14de5e1597048d7e257e
Reviewed-on: https://gerrit.franz.com:9080/8554
Reviewed-by: Martin Mikelsons <mm@franz.com>
Reviewed-by: Kevin Layer <layer@franz.com>
Tested-by: Kevin Layer <layer@franz.com>