bug25008: Request content-type charset parsing was too strict
To determine the character encoding used in a request, function
net.aserve::request-character-encoding looks at the presence of "charset="
in the "Content-Type" request header.

It used to be too strict, e.g. the following header:

  Content-Type: text/plain;charset="utf-8"

was not accepted due to the quotes around "utf-8" and the lack of space
after the ";".

Test cases added.

Change-Id: Idb68d80673f1164d2107a703daabcf91e8e871f1
Reviewed-on: https://gerrit.franz.com:9080/9170
Reviewed-by: Willem Broekema <metawilm@franz.com>
Reviewed-by: Tadeusz Sznuk <ts@franz.com>
Reviewed-by: Martin Mikelsons <mm@franz.com>
Reviewed-by: John Foderaro <jkf@franz.com>
Reviewed-by: Kevin Layer <layer@franz.com>
Tested-by: Kevin Layer <layer@franz.com>