force-output of a prepend-stream
%Testing: aserve/test/t-aserve.cl

The prepend-stream is a stream defined in AllegroServe.
In this change we cause a force-output of a prepend-stream
to pass the data in the prepend-stream's buffer to the underlying
stream and then do a force-output of the underlying stream.
As the prepend-stream is used when sending a response back to
the client this allows a function handling an http request
to send the response back in pieces and have those pieces available
to the client as soon as they are computed.

Change-Id: I92c1d65a479ef7e9ab4945a009bf0a343c99120c
Reviewed-on: https://gerrit.franz.com:9080/9692
Reviewed-by: John Foderaro <jkf@franz.com>
Reviewed-by: Kevin Layer <layer@franz.com>
Tested-by: Kevin Layer <layer@franz.com>