bug25163: Fix setting response trailers when :xmit-server-response-body debug active
Setting trailers on a chunked stream would not work if debug option
:xmit-server-response-body is enabled. That debug setting would
enclose the chunking stream in a broadcast stream, and SET-TRAILERS
has no corresponding method. The result is that the trailers would
be silently dropped.

Fixed by adding a method for SET-TRAILERS and CAN-SET-TRAILERS-P
for broadcast streams.

Let SET-TRAILERS return whether setting all headers succeeds,
this is useful when tracing the function.

New test case added.

Change-Id: I78cebb9f92566f6d0c8e7fdcade12e7be053ceed
Reviewed-on: https://gerrit.franz.com:9080/10035
Reviewed-by: Willem Broekema <metawilm@franz.com>
Reviewed-by: Ahmon Dancy <dancy@franz.com>
Reviewed-by: Martin Mikelsons <mm@franz.com>
Reviewed-by: John Foderaro <jkf@franz.com>
Reviewed-by: Kevin Layer <layer@franz.com>
Tested-by: Kevin Layer <layer@franz.com>