rfe15456: read past the body of get (or any) request
%Testing: aserve/test/t-aserve

According to HTTP spec, any request is allowed to have a body.
We must read past this body if the request handler failed to
do so in order to position an open socket at the correct
beginning of a subsequent request.

Proxies also must forward body of any request if present.

Add tests for new behavior.

Re-indent comments.

Change-Id: Ib16f6be11d39d8fb05263ca9ddb9b3003c220b35
Reviewed-on: https://gerrit.franz.com:9080/10353
Reviewed-by: Martin Mikelsons <mm@franz.com>
Reviewed-by: Duane Rettig <duane@franz.com>
Reviewed-by: John Foderaro <jkf@franz.com>
Reviewed-by: Kevin Layer <layer@franz.com>
Tested-by: Kevin Layer <layer@franz.com>