rfe14988: Add client support for SNI
net.client.aserve:do-http-request now has automatic support for
Service Name Indication (SNI).  The calls to
acl-socket:make-ssl-client-stream will now set the :server-name
keyword value from the host of the URI passed to do-http-request.

This feature of :ssl is only used if it is present at runtime.  The
feature was added as a patch to ACL 10.1.

%Testing: t-aserve, much manual testing

Change-Id: I53cfcbf9a3caa81a92a5bced60a0aad23a320fd9
Reviewed-on: https://gerrit.franz.com:9080/10890
Tested-by: Kevin Layer <layer@franz.com>
Reviewed-by: Duane Rettig <duane@franz.com>
Reviewed-by: Ahmon Dancy <dancy@franz.com>
Reviewed-by: Martin Mikelsons <mm@franz.com>
Reviewed-by: Kevin Layer <layer@franz.com>