rfe15781: webactions 1.18: can use only cookies for sessions
Add a :session-cookie-only argument to webaction-project to
cause Webactions to use cookies exclusively for sessions
of that project and never put the session ID in the url.

%Testing: added test to t-webactions

<release-note>
Add a :session-cookie-only argument to webaction-project to
cause Webactions to use cookies exclusively for sessions
and never put the session ID in the url.
</release-note>

<documentation>
add to the document of webaction-project a new keyword argument
session-cookie-only

with the meaning:  when true will cause Webactions to assume that
a session can be tracked with a cookie and thus it will never put the
session ID in the url.    If the assumption is wrong and a
client does not store the webactions session cookie then sessions
will no work with that client.
</documentation>

Change-Id: Ibac8bbcb2fa5037d01b4a56e532d0473a04a6d0f
Reviewed-on: https://gerrit.franz.com:9080/11467
Reviewed-by: John Foderaro <jkf@franz.com>
Reviewed-by: Kevin Layer <layer@franz.com>
Tested-by: Kevin Layer <layer@franz.com>