bug25472: 1.3.64: proxying https through a tunnel
When proxying an https request do-http-request will now contact the proxy
and establish a tunnel to the target server.

Also some work was done on the test suite to eliminate errors that prevented
it from running to completion.

%Testing: This can't be tested using aserve's own proxy as it doesn't
support tunneling (the CONNECT command).  This change can be tested
using the Fiddler proxy that is designed to help debug proxying.
(see https://www.telerik.com/fiddler).    Fiddler installs
easily on Windows.

<release-note>
bug25472
do-http-request can now proxy https through a proxy that supports
tunnelling (the CONNECT command).
</release-note>

<documentation>
bug25472
We never mentioned that proxying did not work for https so there's
nothing to change in the documentation.
</documentation>

Change-Id: I8c74afbd5c7de72dfe79d682a5fe7ed3ed919ac8
Reviewed-on: https://gerrit.franz.com:9080/11580
Reviewed-by: John Foderaro <jkf@franz.com>
Reviewed-by: Kevin Layer <layer@franz.com>
Tested-by: Kevin Layer <layer@franz.com>