bug25477: device-read for truncated-stream sometimes reads data all the way up to the end
Device-read method for truncated-stream was written in a way that
when a buffer had bigger length than bytes remaining in the stream,
the end parameter was set to read all of the remaining bytes.

This has been corrected to also take into account the initial end
parameter value of device-read.

Change-Id: I84f069a60f47d9fd4014a1fc52fe48d38492d22c
Reviewed-on: https://gerrit.franz.com:9080/11609
Reviewed-by: Bogdan Shcherbyna <bsch@franz.com>
Reviewed-by: Martin Mikelsons <mm@franz.com>
Reviewed-by: Willem Broekema <metawilm@franz.com>
Reviewed-by: Ahmon Dancy <dancy@franz.com>
Reviewed-by: John Foderaro <jkf@franz.com>
Reviewed-by: Kevin Layer <layer@franz.com>
Tested-by: Kevin Layer <layer@franz.com>