rfe15820: 1.3.63:  do request timing by microsecond
As well as recording the universal time when a request
begins being handled and finishes being handled,
we now record also the time in microseconds when it begins and
finishes.  The allows for more accurate logging of the
time taken by a request.  The microsecond time is to be used
only to measure durations, not as a way of specifying
the absolute time.

%Testing:  I wrote some custom logging to show that these values
were being recorded correctly.

<release-note>
rfe15820 1.3.63
Two new fields added to the request object:

request-microtime  with accessor request-request-microtime
reply-microtime    with accessor request-reply-microtime

These record the time in microseconds just before a request
is handled and just after it finishes being handled.
Custom logging functions can use this information to log the
time taken to handle a request.
The microsecond time is to be used only to measure duration, not
as a way of specifying the absolute time.
</release-note>

<documentation>
rfe15820 1.3.63
see release note for information to be added to the aserve manual
</documentation>

Change-Id: I3be649467806b4eddaf4b4d7ff0bcf42f84126a5
Reviewed-on: https://gerrit.franz.com:9080/11737
Reviewed-by: John Foderaro <jkf@franz.com>
Reviewed-by: Ahmon Dancy <dancy@franz.com>
Reviewed-by: Kevin Layer <layer@franz.com>
Tested-by: Kevin Layer <layer@franz.com>