rfe15841: 1.3.66: caching for do-http-request
do-http-request can now work with a cache to reduce network
traffic in computing responses to http requests.

%Bootstrapping: rm -f aserve/*.fasl && make
%Testing: cd aserve; make test

<release-note>
rfe15841
Defined a client-cache object that works with do-http-request to
save HTTP responses and in the future use those saved responses
rather than go out to the network to fetch responses.
</release-note>

<documentation>
rfe15841

Client caching in AllegroServe.

The do-http-request function will cache responses if a
client-cache object is passed as the value of the newly added  :cache argument.

Only :get and :head requests meeting certain criteria are cacheable:
The request cannot have a query, basic or digest authentication, or new headers
passed in.

You can specify the maximum size of the data in the cache by passing the :max-cache-size argument to make-instance of client-cache
e.g.
  (make-instance 'client-cache :max-cache-size 1000000)

When the number of bytes stored in the cache exceeds the max-cache-size
plus *cache-size-slop* (default 100000) entries in the cache are removed
so that the size is less than the maximum size minus *cache-size-slop*.
The least recently used cache entries are removed first.

Accessors for the client-cache object (in the net.aserve.client package):

client-cache-max-cache-entry-size -  if a response to be cached exceeds this
  number of bytes the response will not be cached (default 1,000,000 bytes).

client-cache-max-cache-size  - when the cache size exceeds this value
  plus *cache-size-slop* the cache is automatically reduced in size.

client-cache-lookups - the number of times the cache was checked to see
  if it contained the uri passed to do-http-request

client-cache-alive - the number of times the cache contains the desired
  response and the cache entry hadn't expired.
  In this case the response is returned from the cache.

client-cache-revalidate - the number of times the cache contains the desired
  uri but the cache entry had expired so the cache entry
  must be revalidated

client-cache-validated - the number of times the cache contains the desired
  uri but the cache entry had expired so it was revalidated and found
  to be still valid.

functions:

 do-http-request takes an additional keyword argument :cache which if
supplied must have a client-cache object as a value.  The client cache
object can be shared among multiple threads calling do-http-request to
create a shared cache.

(net.aserve.client:flush-client-cache cache &key expired all)
 Specify which entries in the cache should be removed immediately.
Passing :all a true value will remove all entries.
Passing :expired a true value will remove entries whose expiration
date is in the past.  Note that just because a cache entry has
expired it doesn't mean it's worthless.   When an expired entry
is encountered AllegroServe will do a revalidation request which
may return an indication that the entry is still valid thus saving
the cost of returning the body of the request again.   Also the
cache entry will get a new expiration date.

variable
net.aserve.client:*cache-size-slop*  - specifies how many more bytes
beyond the (client-cache-max-cache-size cache) value are permitted before
the cache size is reduced to *cache-size-slop* less than
the (client-cache-max-cache-size cache) value.

</documentation>

Change-Id: Iaf3decd4de547d5b3dbc721a1399cb415363d834
Reviewed-on: https://gerrit.franz.com:9080/11862
Reviewed-by: John Foderaro <jkf@franz.com>
Reviewed-by: Kevin Layer <layer@franz.com>
Tested-by: Kevin Layer <layer@franz.com>